Iterative deepening done!
Still need to look over why it loses against bad AI when playing as player 2. It seems to beat the random AI as much as it loses against it.
I had a state earlier today when the AI won in all cases but i didn't commit that and i don't know what i did to make it start lose again.